requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
cssselect2==0.8.0
distro==1.9.0
filelock==3.18.0
Flask==3.1.0
fonttools==4.57.0
fsspec==2025.3.2
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
openai==1.75.0
packaging==25.0
pandas==2.2.3
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
pluggy==1.5.0
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydyf==0.11.0
PyPDF2==3.0.1
pypdfium2==4.30.1
pyphen==0.17.2
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
sympy==1.13.3
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.1
torch==2.2.2
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.1.3
zopfli==0.2.3.post1
